new: Add patches to avoid CoW behavior of light arrays
This showed up as a massive hot spot while profiling as these large data
structures were being constantly cloned, placing significant stress on
the garbage collector and leading to a lot of wasted CPU time.

The patch makes use of a double-buffer approach behind a SeqLock to synchronize
the light array views between the game server and light engine actor,
eliminating almost all the copies.

This patch is loosely inspired from another similar one in the Paper project,
but the underlying implementations are significantly different.